requirements.txt
appdirs==1.4.4
atpublic==2.3
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
decorator==4.4.2
dictdiffer==0.8.1
diskcache==5.2.1
distlib==0.3.2
distro==1.5.0
dpath==2.0.1
dulwich==0.20.23
dvc==2.3.0
filelock==3.0.12
Flask==2.0.1
flatten-dict==0.4.0
flufl.lock==3.2
fsspec==2021.5.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.7
GitPython==3.1.17
grandalf==0.6
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpath-ng==1.5.2
mailchecker==4.0.8
MarkupSafe==2.0.1
nanotime==0.5.2
networkx==2.5.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathspec==0.8.1
phonenumbers==8.12.24
pluggy==0.13.1
ply==3.11
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydot==1.4.2
pygit2==1.6.0
Pygments==2.9.0
pygtrie==2.3.2
pyparsing==2.4.7
pytest==6.2.4
python-benedict==0.24.0
python-dateutil==2.8.1
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rich==10.3.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
shortuuid==1.0.1
shtab==1.3.6
six==1.16.0
sklearn==0.0
smmap==4.0.0
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tox==3.23.1
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
voluptuous==0.12.1
wcwidth==0.2.5
Werkzeug==2.0.1
xgboost==1.4.2
xmltodict==0.12.0
zc.lockfile==2.0
